// src/features/about/types.tsx
export interface Info {
  company: {
    ceo: string;
    coo: string;
    cto_propulsion: string;
    cto: string;
    employees: number;
    founded: number;
    founder: string;
    launch_sites: number;
    name: string;
    summary: string;
    test_sites: number;
    valuation: number;
    vehicles: number;
    links: {
      elon_twitter: string;
      flickr: string;
      twitter: string;
      website: string;
    };
    headquarters: {
      address: string;
      city: string;
      state: string;
    };
  };
}